Extract the Address and fix some its minor issues
